temp commit

+++ llarp/dns/srv_data.cpp
@@ -117,7 +117,8 @@ namespace llarp::dns
     if (not bencode_discard(buf))
       return false;
     byte_t* end = buf->cur;
-    std::string_view srvString{reinterpret_cast<char*>(begin), end - begin};
+    std::string_view srvString{
+        reinterpret_cast<char*>(begin), static_cast<std::size_t>(end - begin)};
     try
     {
       SRVTuple tuple{};
